Fix login bug, use fork of funcaptcha, modulerize captcha logic, and more stuff
