+++ package.json
@@ -26,7 +26,7 @@
     "discord.js": "^14.11.0",
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
-    "funcaptcha": "^1.0.37",
+    "funcaptcha": "github:AlexandraBaker/funcaptcha#patch-8",
     "ms": "^2.1.3",
     "noblox.js": "^4.14.1"
   },
